old2/requirements.txt
pandas
numpy
openpyxl
websocket-client
websockets
xgboost
scikit-learn
joblib
ta-lib
matplotlib
seaborn
